requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
pymysql
cryptography
requests
python-dotenv
reportlab
gunicorn